+++ lib/impaktfull_ui.dart
@@ -132,6 +132,7 @@ export 'src/util/extension/text_style_extension.dart';
 
 /// Util
 export 'src/util/responsive/responsive_extension.dart';
+export 'src/util/localizations/localizations.dart';
 
 /// External packages
 export 'package:snacky/snacky.dart';